Move ConsentUtils to its own file
