+++ src/lib/__init__.py
@@ -1,4 +1,5 @@
 from .code_schemes import CodeSchemes
+from .consent_utils import ConsentUtils
 from .icr_tools import ICRTools
 from .message_filters import MessageFilters
 from .pipeline_configuration import PipelineConfiguration
